Added fix replacing NaN values in primary photo with 'default' string allowing MapBox to render now
